refactor(reducers): move `selected-index`
