+++ app/renderer/src/reducers/selected-index/index.js
@@ -1,0 +1,1 @@
+export { default } from './selected-index';
